add new endpoint for update userAccountType
