+++ src/main/java/com/school/project/controller/UserController.java
@@ -2,7 +2,6 @@ package com.school.project.controller;
 
 import com.school.project.dto.UserDto;
 import com.school.project.model.entities.User;
-import com.school.project.model.types.UserAccountType;
 import com.school.project.service.UserAccountService;
 import com.school.project.service.UserService;
 import org.modelmapper.ModelMapper;
@@ -33,22 +32,6 @@ public class UserController {
                         .create(convertUserDtoToUser(userDto))));
     }
 
-    @PutMapping("/users/userAccounts/{id}")
-    public ResponseEntity updateUserAccountType(@PathVariable long id, @RequestBody int accountType){
-        userAccountService.getUserAccountByUserId(id).setAccountRole(chooseType(accountType));
-        return ResponseEntity.ok().build();
-    }
-
-    private UserAccountType chooseType(int accountType){
-        if(accountType==1){
-            return UserAccountType.TEACHER;
-        }
-        if(accountType==2){
-            return UserAccountType.ADMIN;
-        }
-        return UserAccountType.STUDENT;
-    }
-
     @GetMapping("/users/{id}")
     public ResponseEntity<UserDto> getUserByID(@PathVariable long id) {
         return ResponseEntity.ok()
@@ -66,6 +49,12 @@ public class UserController {
     @PutMapping("/users/{id}")
     public ResponseEntity updateUser(@RequestBody UserDto userDto, @PathVariable Long id) {
         userService.update(convertUserDtoToUser(userDto), id);
+        return ResponseEntity.ok().build();
+    }
+
+    @PutMapping("/users/{userId}/status/{statusId}")
+    public ResponseEntity updateAccountRole(@PathVariable Long userId, @PathVariable Integer statusId){
+        userService.updateAccountRole(userId, statusId);
         return ResponseEntity.ok().build();
     }
