+++ src/main/java/com/school/project/service/UserService.java
@@ -1,6 +1,7 @@
 package com.school.project.service;
 
 import com.school.project.model.entities.User;
+import com.school.project.model.entities.UserAccount;
 
 import java.util.List;
 
@@ -15,5 +16,9 @@ public interface UserService {
     User getUserById(Long id);
 
     List<User> getAllUsers();
+
+    void updateAccountRole(Long userId, Integer statusId);
+
+    UserAccount getUserAccountByUserId(Long id);
 
 }
