+++ src/main/java/com/school/project/service/impl/UserAccountServiceImpl.java
@@ -1,6 +1,5 @@
 package com.school.project.service.impl;
 
-import com.school.project.exception.UserAccountNotFoundException;
 import com.school.project.model.entities.User;
 import com.school.project.model.entities.UserAccount;
 import com.school.project.model.types.UserAccountType;
@@ -8,8 +7,6 @@ import com.school.project.repository.UserAccountRepository;
 import com.school.project.service.UserAccountService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import java.util.Optional;
 
 @Service
 public class UserAccountServiceImpl implements UserAccountService {
@@ -25,20 +22,6 @@ public class UserAccountServiceImpl implements UserAccountService {
                 .build();
         userAccountRepository.save(userAccount);
         return userAccount;
-    }
-
-    @Override
-    public UserAccount getUserAccountById(Long id) {
-        Optional<UserAccount> byId = userAccountRepository.findById(id);
-        if (!byId.isPresent()) throw new UserAccountNotFoundException("User account with id : " + id + " is not found");
-        return byId.get();
-    }
-
-    @Override
-    public UserAccount getUserAccountByUserId(Long id) {
-        Optional<UserAccount> userAccountByUserId = userAccountRepository.getUserAccountByUserId(id);
-        if (!userAccountByUserId.isPresent()) throw new UserAccountNotFoundException("User account with id : " + id + " is not found");
-        return userAccountByUserId.get();
     }
 
 }
